CargoStore: Re-check mandatory fields after unique / regex constraints
A field value may be blanked as a result of violating the field's uniqueness or
regex constraint; the blanked value raises a SQL error if the field is also
non-null. This patch modifies CargoStore::blankOrRejectBadData so that the
mandatory constraint of the field is checked again if the value is blanked.

The mandatory constraint can also simply be checked after all other constraints,
but doing it this way provides slightly more descriptive error messages.

Bug: T234443
Change-Id: I38b416540a892b3973400cb9650cfc180914adb2