viewMoreResultsLink(): Pass "group_by" and "having" verbatim
The CargoSQLQuery class transforms the group-by clause from Cargo's format to
SQL's; since the Special:CargoQuery page does it anyway, there is no need for
the "more results" links to repeat the transformation in the URL parameters
(which in some cases also exposes the underlying table identifiers).

The same applies to the having clause, but apparently the clause was never
added to "more results" links in the first place, so this patch adds that as
well.

Bug: T235310
Change-Id: Ifbeaf2a6ba391d1093e5df10105e4316c605cd37