Account for non-UTC $wgLocaltimezone in icalendar exports
Page values such as start and end dates are stored in the wiki's
timezone, so need to be converted to UTC when put into icalendar.
The page modification date is stored in UTC, so does not need to
be.

This adds a test for a different wiki timezone and fixes the start
and end dates of events.

Bug: T250547
Change-Id: Ib70bb6118c40ef66858f0ea1db011a37326ce792