GPUCP-356 | Revert warning box removal for hydra branch
In current gamepedia backend warning box works correctly, so it should be turned on until all of gampedias will be migrated into Fandom backend. For UCP we will be using `ucp` branch were this message will be hidden until we will have solution for how to show this message correctly for the users. It is currently impossible in our backend to get those information because we are using Rabbitmq as message broker which doesn't provide Api to get information about current tasks in queue.