Rework Cargo data storage
Instead of trying to switch whether cargo_store actually does writes in
the parserfunction itself, always do the writes, saving what would be
written to ParserOutput extensionData, and rollback at the end of the
parse.

Then, in LinksUpdate (instead of PageContentSaveComplete), perform the
same writes to cargo tables and commit the writes.

This leads to all parses having a consistent view of data they store,
meaning previewing edits in situations where a page stores data and then
queries that data appears to act logically now.

On the other hand, this means that if the LinksUpdate writes fail, a
page can have data in cargo tables which differs from what appears to
show on the page.  This is a rare case, however, and it's one I'm more
comfortable with than the current status quo of "no data saves until you
save the page" and "parses (almost) always look at old data".